pyproject.toml
[project]
name = "loulett-bot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "python-telegram-bot>=22.6",
]
